WIP: Refactor old PDF module to a new package
* Adds tests.
