+++ packages/pdf/src/RPA/PDF/__init__.py
@@ -55,6 +55,10 @@ class PDF(DynamicCore):
     def __init__(self):
         self.logger = logging.getLogger(__name__)
         self.buffer = Buffer(self.logger)
+        self.active_fields = None
+        self.active_fileobject = None
+        self.active_pdf = None
+        self.fileobjects = {}
         self.rpa_pdf_document = None
 
         # Register keyword libraries to LibCore
@@ -64,3 +68,22 @@ class PDF(DynamicCore):
             ModelKeywords(self),
         ]
         super().__init__(libraries)
+
+        # TODO: how to use this RPA.main keyword library
+        # listener = RobotLogListener()
+        # listener.register_protected_keywords(["RPA.PDF.decrypt"])
+
+    def __del__(self):
+        self.close_all_pdf_documents()
+
+    def close_all_pdf_documents(self) -> None:
+        """Close all opened PDF file descriptors."""
+        for filename, fileobject in self.fileobjects.items():
+            fileobject.close()
+            self.logger.debug('PDF "%s" closed', filename)
+        self.active_fields = None
+        self.active_pdf = None
+        self.active_fileobject = None
+        self.anchor_element = None
+        self.fileobjects = {}
+        self.rpa_pdf_document = None
