+++ packages/pdf/tests/resources/imagesandtext.pdf
