+++ packages/pdf/tests/resources/vero.pdf
